Update App.js
Add the navigation to mock exam progress resources suggestions. The resources page was designed to be the video page. I change the file name and menu item name in topnav but this link still works. If the link doesnt work, Ill fix it but for now just leave it as it is.